refactored comparison logic to make more modular and successfully removed double call to api in the process
