+++ assets/top20.json
@@ -1,1 +1,1 @@
-[{"id":97928,"appid":"730","name":"Counter-Strike: Global Offensive","isgame":1},{"id":149219,"appid":"570","name":"Dota 2","isgame":1},{"id":14451,"appid":"578080","name":"PUBG: BATTLEGROUNDS","isgame":1},{"id":146483,"appid":"1172470","name":"Apex Legends","isgame":1},{"id":146903,"appid":"1599340","name":"Lost Ark","isgame":1},{"id":147642,"appid":"271590","name":"Grand Theft Auto V","isgame":1},{"id":150862,"appid":"1203220","name":"NARAKA: BLADEPOINT","isgame":1},{"id":97967,"appid":"440","name":"Team Fortress 2","isgame":1},{"id":150657,"appid":"431960","name":"Wallpaper Engine","isgame":1},{"id":148014,"appid":"1091500","name":"Cyberpunk 2077","isgame":1},{"id":91683,"appid":"105600","name":"Terraria","isgame":1},{"id":88394,"appid":"252490","name":"Rust","isgame":1},{"id":148390,"appid":"1085660","name":"Destiny 2","isgame":1},{"id":2334,"appid":"1222680","name":"Need for Speed™ Heat ","isgame":1},{"id":120767,"appid":"346110","name":"ARK: Survival Evolved","isgame":1},{"id":148137,"appid":"218620","name":"PAYDAY 2","isgame":1},{"id":150493,"appid":"236390","name":"War Thunder","isgame":1},{"id":112314,"appid":"394360","name":"Hearts of Iron IV","isgame":1},{"id":148864,"appid":"289070","name":"Sid Meier's Civilization VI","isgame":1},{"id":149590,"appid":"1811260","name":"EA SPORTS™ FIFA 23","isgame":1}]
+[{"id":97928,"appid":"730","name":"Counter-Strike: Global Offensive","isgame":1},{"id":149219,"appid":"570","name":"Dota 2","isgame":1},{"id":146483,"appid":"1172470","name":"Apex Legends","isgame":1},{"id":146903,"appid":"1599340","name":"Lost Ark","isgame":1},{"id":14451,"appid":"578080","name":"PUBG: BATTLEGROUNDS","isgame":1},{"id":97967,"appid":"440","name":"Team Fortress 2","isgame":1},{"id":148014,"appid":"1091500","name":"Cyberpunk 2077","isgame":1},{"id":148390,"appid":"1085660","name":"Destiny 2","isgame":1},{"id":147642,"appid":"271590","name":"Grand Theft Auto V","isgame":1},{"id":91683,"appid":"105600","name":"Terraria","isgame":1},{"id":150657,"appid":"431960","name":"Wallpaper Engine","isgame":1},{"id":150862,"appid":"1203220","name":"NARAKA: BLADEPOINT","isgame":1},{"id":88394,"appid":"252490","name":"Rust","isgame":1},{"id":2334,"appid":"1222680","name":"Need for Speed™ Heat ","isgame":1},{"id":148137,"appid":"218620","name":"PAYDAY 2","isgame":1},{"id":120767,"appid":"346110","name":"ARK: Survival Evolved","isgame":1},{"id":121093,"appid":"1623660","name":"MIR4","isgame":1},{"id":148864,"appid":"289070","name":"Sid Meier's Civilization VI","isgame":1},{"id":146949,"appid":"230410","name":"Warframe","isgame":1},{"id":150493,"appid":"236390","name":"War Thunder","isgame":1}]
